split off voting buttons
